Make `subs` and `hdr` optional for `@subroutine`, and make `compute_post_command_offsets` and `update_commands_with_offsets` take a `manager` parameter rather than use `CommonGlobals.script_manager`

+++ mnlscript/dt/commands.py
@@ -526,7 +526,7 @@ def get_actor_attribute(
 def set_actor_attribute(
     actor: int | mnllib.Variable,
     attribute: ActorAttribute | int | mnllib.Variable,
-    value: int | mnllib.Variable,
+    value: float | mnllib.Variable,
     *,
     subroutine: mnllib.Subroutine | None = None,
 ) -> mnllib.CodeCommand:
